+++ yash-env/CHANGELOG.md
@@ -13,7 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The `builtin::Builtin` struct now has the `is_declaration_utility` field.
 - The `builtin::Builtin` struct now can be constructed with the associated
   function `new`.
+- The `system::errno::Errno` struct now can be converted to and from the `Errno`
+  type from the `errno` crate.
 - Internal dependencies:
+    - errno 0.3.10
     - libc 0.2.169
 
 ### Changed